City Battle - Census Data
The problem faced by many people is where to live and whether they will find work in that location. This code allows a user to select characteristics about themselves and identify which city is best for them to find a job. The user is prompted for their sex and age and is then provided with a sorted list of the total job hires for peoples matching their characteristics. The data referenced is the government census data which is a reliable and free data source.